+++ src/jdk.incubator.vector/share/classes/jdk/incubator/vector/ByteVector.java
@@ -3460,7 +3460,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
             this,
             a, offset,
             (arr, off, v)
-            -> v.stOp(arr, off,
+            -> v.stOp(arr, (int) off,
                       (arr_, off_, i, e) -> arr_[off_ + i] = e));
     }
 
@@ -3611,7 +3611,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
             normalized,
             a, offset,
             (arr, off, v)
-            -> v.stOp(arr, off,
+            -> v.stOp(arr, (int) off,
                       (arr_, off_, i, e) -> arr_[off_ + i] = (e & 1) != 0));
     }
 
@@ -3881,7 +3881,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
             vsp.vectorType(), vsp.elementType(), vsp.laneCount(),
             a, arrayAddress(a, offset),
             a, offset, vsp,
-            (arr, off, s) -> s.ldOp(arr, off,
+            (arr, off, s) -> s.ldOp(arr, (int) off,
                                     (arr_, off_, i) -> arr_[off_ + i]));
     }
 
@@ -3898,7 +3898,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
             vsp.vectorType(), maskClass, vsp.elementType(), vsp.laneCount(),
             a, arrayAddress(a, offset), m,
             a, offset, vsp,
-            (arr, off, s, vm) -> s.ldOp(arr, off, vm,
+            (arr, off, s, vm) -> s.ldOp(arr, (int) off, vm,
                                         (arr_, off_, i) -> arr_[off_ + i]));
     }
 
@@ -3915,7 +3915,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
             vsp.vectorType(), vsp.elementType(), vsp.laneCount(),
             a, booleanArrayAddress(a, offset),
             a, offset, vsp,
-            (arr, off, s) -> s.ldOp(arr, off,
+            (arr, off, s) -> s.ldOp(arr, (int) off,
                                     (arr_, off_, i) -> (byte) (arr_[off_ + i] ? 1 : 0)));
     }
 
@@ -3932,7 +3932,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
             vsp.vectorType(), maskClass, vsp.elementType(), vsp.laneCount(),
             a, booleanArrayAddress(a, offset), m,
             a, offset, vsp,
-            (arr, off, s, vm) -> s.ldOp(arr, off, vm,
+            (arr, off, s, vm) -> s.ldOp(arr, (int) off, vm,
                                         (arr_, off_, i) -> (byte) (arr_[off_ + i] ? 1 : 0)));
     }
 
@@ -3949,7 +3949,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
             a, offset, vsp,
             (arr, off, s) -> {
                 ByteBuffer wb = wrapper(arr, NATIVE_ENDIAN);
-                return s.ldOp(wb, off,
+                return s.ldOp(wb, (int) off,
                         (wb_, o, i) -> wb_.get(o + i * 1));
             });
     }
@@ -3968,7 +3968,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
             a, offset, vsp,
             (arr, off, s, vm) -> {
                 ByteBuffer wb = wrapper(arr, NATIVE_ENDIAN);
-                return s.ldOp(wb, off, vm,
+                return s.ldOp(wb, (int) off, vm,
                         (wb_, o, i) -> wb_.get(o + i * 1));
             });
     }
@@ -3984,7 +3984,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
                 bb, offset, vsp,
                 (buf, off, s) -> {
                     ByteBuffer wb = wrapper(buf, NATIVE_ENDIAN);
-                    return s.ldOp(wb, off,
+                    return s.ldOp(wb, (int) off,
                             (wb_, o, i) -> wb_.get(o + i * 1));
                 });
     }
@@ -4002,7 +4002,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
                 bb, offset, m, vsp,
                 (buf, off, s, vm) -> {
                     ByteBuffer wb = wrapper(buf, NATIVE_ENDIAN);
-                    return s.ldOp(wb, off, vm,
+                    return s.ldOp(wb, (int) off, vm,
                             (wb_, o, i) -> wb_.get(o + i * 1));
                 });
     }
@@ -4015,10 +4015,10 @@ public abstract class ByteVector extends AbstractVector<Byte> {
         ByteSpecies vsp = vspecies();
         return ScopedMemoryAccess.loadFromMemorySegment(
                 vsp.vectorType(), vsp.elementType(), vsp.laneCount(),
-                (MemorySegmentProxy) ms, (int) offset, vsp, // @@@ downcast from long to int
+                (MemorySegmentProxy) ms, offset, vsp,
                 (msp, off, s) -> {
                     var layout = ValueLayout.JAVA_BYTE.withBitAlignment(8);
-                    return s.ldOp((MemorySegment) msp, off,
+                    return s.ldOp((MemorySegment) msp, (int) off, // @@@ downcast from long to int
                             (ms_, o, i) -> ms_.get(layout, o + i * 1L));
                 });
     }
@@ -4033,10 +4033,10 @@ public abstract class ByteVector extends AbstractVector<Byte> {
         m.check(vsp);
         return ScopedMemoryAccess.loadFromMemorySegmentMasked(
                 vsp.vectorType(), maskClass, vsp.elementType(), vsp.laneCount(),
-                (MemorySegmentProxy) ms, (int) offset, m, vsp, // @@@ downcast from long to int
+                (MemorySegmentProxy) ms, offset, m, vsp,
                 (msp, off, s, vm) -> {
                     var layout = ValueLayout.JAVA_BYTE.withBitAlignment(8);
-                    return s.ldOp((MemorySegment) msp, off, vm,
+                    return s.ldOp((MemorySegment) msp, (int) off, vm, // @@@ downcast from long to int
                             (ms_, o, i) -> ms_.get(layout, o + i * 1L));
                 });
     }
@@ -4056,7 +4056,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
             a, arrayAddress(a, offset),
             this, a, offset,
             (arr, off, v)
-            -> v.stOp(arr, off,
+            -> v.stOp(arr, (int) off,
                       (arr_, off_, i, e) -> arr_[off_+i] = e));
     }
 
@@ -4073,7 +4073,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
             a, arrayAddress(a, offset),
             this, m, a, offset,
             (arr, off, v, vm)
-            -> v.stOp(arr, off, vm,
+            -> v.stOp(arr, (int) off, vm,
                       (arr_, off_, i, e) -> arr_[off_ + i] = e));
     }
 
@@ -4092,7 +4092,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
             a, booleanArrayAddress(a, offset),
             normalized, m, a, offset,
             (arr, off, v, vm)
-            -> v.stOp(arr, off, vm,
+            -> v.stOp(arr, (int) off, vm,
                       (arr_, off_, i, e) -> arr_[off_ + i] = (e & 1) != 0));
     }
 
@@ -4108,7 +4108,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
             this, a, offset,
             (arr, off, v) -> {
                 ByteBuffer wb = wrapper(arr, NATIVE_ENDIAN);
-                v.stOp(wb, off,
+                v.stOp(wb, (int) off,
                         (tb_, o, i, e) -> tb_.put(o + i * 1, e));
             });
     }
@@ -4127,7 +4127,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
             this, m, a, offset,
             (arr, off, v, vm) -> {
                 ByteBuffer wb = wrapper(arr, NATIVE_ENDIAN);
-                v.stOp(wb, off, vm,
+                v.stOp(wb, (int) off, vm,
                         (tb_, o, i, e) -> tb_.put(o + i * 1, e));
             });
     }
@@ -4141,7 +4141,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
                 this, bb, offset,
                 (buf, off, v) -> {
                     ByteBuffer wb = wrapper(buf, NATIVE_ENDIAN);
-                    v.stOp(wb, off,
+                    v.stOp(wb, (int) off,
                             (wb_, o, i, e) -> wb_.put(o + i * 1, e));
                 });
     }
@@ -4159,7 +4159,7 @@ public abstract class ByteVector extends AbstractVector<Byte> {
                 this, m, bb, offset,
                 (buf, off, v, vm) -> {
                     ByteBuffer wb = wrapper(buf, NATIVE_ENDIAN);
-                    v.stOp(wb, off, vm,
+                    v.stOp(wb, (int) off, vm,
                             (wb_, o, i, e) -> wb_.put(o + i * 1, e));
                 });
     }
@@ -4171,10 +4171,10 @@ public abstract class ByteVector extends AbstractVector<Byte> {
         ScopedMemoryAccess.storeIntoMemorySegment(
                 vsp.vectorType(), vsp.elementType(), vsp.laneCount(),
                 this,
-                (MemorySegmentProxy) ms, (int) offset, // @@@ downcast from long to int
+                (MemorySegmentProxy) ms, offset,
                 (msp, off, v) -> {
                     var layout = ValueLayout.JAVA_BYTE.withBitAlignment(8);
-                    v.stOp((MemorySegment) msp, off,
+                    v.stOp((MemorySegment) msp, (int) off, // @@@ downcast from long to int
                             (ms_, o, i, e) -> ms_.set(layout, o + i * 1L, e));
                 });
     }
@@ -4190,10 +4190,10 @@ public abstract class ByteVector extends AbstractVector<Byte> {
         ScopedMemoryAccess.storeIntoMemorySegmentMasked(
                 vsp.vectorType(), maskClass, vsp.elementType(), vsp.laneCount(),
                 this, m,
-                (MemorySegmentProxy) ms, (int) offset, // @@@ downcast from long to int
+                (MemorySegmentProxy) ms, offset,
                 (msp, off, v, vm) -> {
                     var layout = ValueLayout.JAVA_BYTE.withBitAlignment(8);
-                    v.stOp((MemorySegment) msp, off, vm,
+                    v.stOp((MemorySegment) msp, (int) off, vm,  // @@@ downcast from long to int
                             (ms_, o, i, e) -> ms_.set(layout, o + i * 1L, e));
                 });
     }
