+++ src/jdk.incubator.vector/share/classes/jdk/incubator/vector/IntVector.java
@@ -3320,7 +3320,7 @@ public abstract class IntVector extends AbstractVector<Integer> {
             this,
             a, offset,
             (arr, off, v)
-            -> v.stOp(arr, off,
+            -> v.stOp(arr, (int) off,
                       (arr_, off_, i, e) -> arr_[off_ + i] = e));
     }
 
@@ -3593,7 +3593,7 @@ public abstract class IntVector extends AbstractVector<Integer> {
             vsp.vectorType(), vsp.elementType(), vsp.laneCount(),
             a, arrayAddress(a, offset),
             a, offset, vsp,
-            (arr, off, s) -> s.ldOp(arr, off,
+            (arr, off, s) -> s.ldOp(arr, (int) off,
                                     (arr_, off_, i) -> arr_[off_ + i]));
     }
 
@@ -3610,7 +3610,7 @@ public abstract class IntVector extends AbstractVector<Integer> {
             vsp.vectorType(), maskClass, vsp.elementType(), vsp.laneCount(),
             a, arrayAddress(a, offset), m,
             a, offset, vsp,
-            (arr, off, s, vm) -> s.ldOp(arr, off, vm,
+            (arr, off, s, vm) -> s.ldOp(arr, (int) off, vm,
                                         (arr_, off_, i) -> arr_[off_ + i]));
     }
 
@@ -3663,7 +3663,7 @@ public abstract class IntVector extends AbstractVector<Integer> {
             a, offset, vsp,
             (arr, off, s) -> {
                 ByteBuffer wb = wrapper(arr, NATIVE_ENDIAN);
-                return s.ldOp(wb, off,
+                return s.ldOp(wb, (int) off,
                         (wb_, o, i) -> wb_.getInt(o + i * 4));
             });
     }
@@ -3682,7 +3682,7 @@ public abstract class IntVector extends AbstractVector<Integer> {
             a, offset, vsp,
             (arr, off, s, vm) -> {
                 ByteBuffer wb = wrapper(arr, NATIVE_ENDIAN);
-                return s.ldOp(wb, off, vm,
+                return s.ldOp(wb, (int) off, vm,
                         (wb_, o, i) -> wb_.getInt(o + i * 4));
             });
     }
@@ -3698,7 +3698,7 @@ public abstract class IntVector extends AbstractVector<Integer> {
                 bb, offset, vsp,
                 (buf, off, s) -> {
                     ByteBuffer wb = wrapper(buf, NATIVE_ENDIAN);
-                    return s.ldOp(wb, off,
+                    return s.ldOp(wb, (int) off,
                             (wb_, o, i) -> wb_.getInt(o + i * 4));
                 });
     }
@@ -3716,7 +3716,7 @@ public abstract class IntVector extends AbstractVector<Integer> {
                 bb, offset, m, vsp,
                 (buf, off, s, vm) -> {
                     ByteBuffer wb = wrapper(buf, NATIVE_ENDIAN);
-                    return s.ldOp(wb, off, vm,
+                    return s.ldOp(wb, (int) off, vm,
                             (wb_, o, i) -> wb_.getInt(o + i * 4));
                 });
     }
@@ -3729,10 +3729,10 @@ public abstract class IntVector extends AbstractVector<Integer> {
         IntSpecies vsp = vspecies();
         return ScopedMemoryAccess.loadFromMemorySegment(
                 vsp.vectorType(), vsp.elementType(), vsp.laneCount(),
-                (MemorySegmentProxy) ms, (int) offset, vsp, // @@@ downcast from long to int
+                (MemorySegmentProxy) ms, offset, vsp,
                 (msp, off, s) -> {
                     var layout = ValueLayout.JAVA_INT.withBitAlignment(8);
-                    return s.ldOp((MemorySegment) msp, off,
+                    return s.ldOp((MemorySegment) msp, (int) off, // @@@ downcast from long to int
                             (ms_, o, i) -> ms_.get(layout, o + i * 4L));
                 });
     }
@@ -3747,10 +3747,10 @@ public abstract class IntVector extends AbstractVector<Integer> {
         m.check(vsp);
         return ScopedMemoryAccess.loadFromMemorySegmentMasked(
                 vsp.vectorType(), maskClass, vsp.elementType(), vsp.laneCount(),
-                (MemorySegmentProxy) ms, (int) offset, m, vsp, // @@@ downcast from long to int
+                (MemorySegmentProxy) ms, offset, m, vsp,
                 (msp, off, s, vm) -> {
                     var layout = ValueLayout.JAVA_INT.withBitAlignment(8);
-                    return s.ldOp((MemorySegment) msp, off, vm,
+                    return s.ldOp((MemorySegment) msp, (int) off, vm, // @@@ downcast from long to int
                             (ms_, o, i) -> ms_.get(layout, o + i * 4L));
                 });
     }
@@ -3770,7 +3770,7 @@ public abstract class IntVector extends AbstractVector<Integer> {
             a, arrayAddress(a, offset),
             this, a, offset,
             (arr, off, v)
-            -> v.stOp(arr, off,
+            -> v.stOp(arr, (int) off,
                       (arr_, off_, i, e) -> arr_[off_+i] = e));
     }
 
@@ -3787,7 +3787,7 @@ public abstract class IntVector extends AbstractVector<Integer> {
             a, arrayAddress(a, offset),
             this, m, a, offset,
             (arr, off, v, vm)
-            -> v.stOp(arr, off, vm,
+            -> v.stOp(arr, (int) off, vm,
                       (arr_, off_, i, e) -> arr_[off_ + i] = e));
     }
 
@@ -3838,7 +3838,7 @@ public abstract class IntVector extends AbstractVector<Integer> {
             this, a, offset,
             (arr, off, v) -> {
                 ByteBuffer wb = wrapper(arr, NATIVE_ENDIAN);
-                v.stOp(wb, off,
+                v.stOp(wb, (int) off,
                         (tb_, o, i, e) -> tb_.putInt(o + i * 4, e));
             });
     }
@@ -3857,7 +3857,7 @@ public abstract class IntVector extends AbstractVector<Integer> {
             this, m, a, offset,
             (arr, off, v, vm) -> {
                 ByteBuffer wb = wrapper(arr, NATIVE_ENDIAN);
-                v.stOp(wb, off, vm,
+                v.stOp(wb, (int) off, vm,
                         (tb_, o, i, e) -> tb_.putInt(o + i * 4, e));
             });
     }
@@ -3871,7 +3871,7 @@ public abstract class IntVector extends AbstractVector<Integer> {
                 this, bb, offset,
                 (buf, off, v) -> {
                     ByteBuffer wb = wrapper(buf, NATIVE_ENDIAN);
-                    v.stOp(wb, off,
+                    v.stOp(wb, (int) off,
                             (wb_, o, i, e) -> wb_.putInt(o + i * 4, e));
                 });
     }
@@ -3889,7 +3889,7 @@ public abstract class IntVector extends AbstractVector<Integer> {
                 this, m, bb, offset,
                 (buf, off, v, vm) -> {
                     ByteBuffer wb = wrapper(buf, NATIVE_ENDIAN);
-                    v.stOp(wb, off, vm,
+                    v.stOp(wb, (int) off, vm,
                             (wb_, o, i, e) -> wb_.putInt(o + i * 4, e));
                 });
     }
@@ -3901,10 +3901,10 @@ public abstract class IntVector extends AbstractVector<Integer> {
         ScopedMemoryAccess.storeIntoMemorySegment(
                 vsp.vectorType(), vsp.elementType(), vsp.laneCount(),
                 this,
-                (MemorySegmentProxy) ms, (int) offset, // @@@ downcast from long to int
+                (MemorySegmentProxy) ms, offset,
                 (msp, off, v) -> {
                     var layout = ValueLayout.JAVA_INT.withBitAlignment(8);
-                    v.stOp((MemorySegment) msp, off,
+                    v.stOp((MemorySegment) msp, (int) off, // @@@ downcast from long to int
                             (ms_, o, i, e) -> ms_.set(layout, o + i * 4L, e));
                 });
     }
@@ -3920,10 +3920,10 @@ public abstract class IntVector extends AbstractVector<Integer> {
         ScopedMemoryAccess.storeIntoMemorySegmentMasked(
                 vsp.vectorType(), maskClass, vsp.elementType(), vsp.laneCount(),
                 this, m,
-                (MemorySegmentProxy) ms, (int) offset, // @@@ downcast from long to int
+                (MemorySegmentProxy) ms, offset,
                 (msp, off, v, vm) -> {
                     var layout = ValueLayout.JAVA_INT.withBitAlignment(8);
-                    v.stOp((MemorySegment) msp, off, vm,
+                    v.stOp((MemorySegment) msp, (int) off, vm,  // @@@ downcast from long to int
                             (ms_, o, i, e) -> ms_.set(layout, o + i * 4L, e));
                 });
     }
