+++ src/jdk.incubator.vector/share/classes/jdk/incubator/vector/LongVector.java
@@ -3199,7 +3199,7 @@ public abstract class LongVector extends AbstractVector<Long> {
             this,
             a, offset,
             (arr, off, v)
-            -> v.stOp(arr, off,
+            -> v.stOp(arr, (int) off,
                       (arr_, off_, i, e) -> arr_[off_ + i] = e));
     }
 
@@ -3491,7 +3491,7 @@ public abstract class LongVector extends AbstractVector<Long> {
             vsp.vectorType(), vsp.elementType(), vsp.laneCount(),
             a, arrayAddress(a, offset),
             a, offset, vsp,
-            (arr, off, s) -> s.ldOp(arr, off,
+            (arr, off, s) -> s.ldOp(arr, (int) off,
                                     (arr_, off_, i) -> arr_[off_ + i]));
     }
 
@@ -3508,7 +3508,7 @@ public abstract class LongVector extends AbstractVector<Long> {
             vsp.vectorType(), maskClass, vsp.elementType(), vsp.laneCount(),
             a, arrayAddress(a, offset), m,
             a, offset, vsp,
-            (arr, off, s, vm) -> s.ldOp(arr, off, vm,
+            (arr, off, s, vm) -> s.ldOp(arr, (int) off, vm,
                                         (arr_, off_, i) -> arr_[off_ + i]));
     }
 
@@ -3579,7 +3579,7 @@ public abstract class LongVector extends AbstractVector<Long> {
             a, offset, vsp,
             (arr, off, s) -> {
                 ByteBuffer wb = wrapper(arr, NATIVE_ENDIAN);
-                return s.ldOp(wb, off,
+                return s.ldOp(wb, (int) off,
                         (wb_, o, i) -> wb_.getLong(o + i * 8));
             });
     }
@@ -3598,7 +3598,7 @@ public abstract class LongVector extends AbstractVector<Long> {
             a, offset, vsp,
             (arr, off, s, vm) -> {
                 ByteBuffer wb = wrapper(arr, NATIVE_ENDIAN);
-                return s.ldOp(wb, off, vm,
+                return s.ldOp(wb, (int) off, vm,
                         (wb_, o, i) -> wb_.getLong(o + i * 8));
             });
     }
@@ -3614,7 +3614,7 @@ public abstract class LongVector extends AbstractVector<Long> {
                 bb, offset, vsp,
                 (buf, off, s) -> {
                     ByteBuffer wb = wrapper(buf, NATIVE_ENDIAN);
-                    return s.ldOp(wb, off,
+                    return s.ldOp(wb, (int) off,
                             (wb_, o, i) -> wb_.getLong(o + i * 8));
                 });
     }
@@ -3632,7 +3632,7 @@ public abstract class LongVector extends AbstractVector<Long> {
                 bb, offset, m, vsp,
                 (buf, off, s, vm) -> {
                     ByteBuffer wb = wrapper(buf, NATIVE_ENDIAN);
-                    return s.ldOp(wb, off, vm,
+                    return s.ldOp(wb, (int) off, vm,
                             (wb_, o, i) -> wb_.getLong(o + i * 8));
                 });
     }
@@ -3645,10 +3645,10 @@ public abstract class LongVector extends AbstractVector<Long> {
         LongSpecies vsp = vspecies();
         return ScopedMemoryAccess.loadFromMemorySegment(
                 vsp.vectorType(), vsp.elementType(), vsp.laneCount(),
-                (MemorySegmentProxy) ms, (int) offset, vsp, // @@@ downcast from long to int
+                (MemorySegmentProxy) ms, offset, vsp,
                 (msp, off, s) -> {
                     var layout = ValueLayout.JAVA_LONG.withBitAlignment(8);
-                    return s.ldOp((MemorySegment) msp, off,
+                    return s.ldOp((MemorySegment) msp, (int) off, // @@@ downcast from long to int
                             (ms_, o, i) -> ms_.get(layout, o + i * 8L));
                 });
     }
@@ -3663,10 +3663,10 @@ public abstract class LongVector extends AbstractVector<Long> {
         m.check(vsp);
         return ScopedMemoryAccess.loadFromMemorySegmentMasked(
                 vsp.vectorType(), maskClass, vsp.elementType(), vsp.laneCount(),
-                (MemorySegmentProxy) ms, (int) offset, m, vsp, // @@@ downcast from long to int
+                (MemorySegmentProxy) ms, offset, m, vsp,
                 (msp, off, s, vm) -> {
                     var layout = ValueLayout.JAVA_LONG.withBitAlignment(8);
-                    return s.ldOp((MemorySegment) msp, off, vm,
+                    return s.ldOp((MemorySegment) msp, (int) off, vm, // @@@ downcast from long to int
                             (ms_, o, i) -> ms_.get(layout, o + i * 8L));
                 });
     }
@@ -3686,7 +3686,7 @@ public abstract class LongVector extends AbstractVector<Long> {
             a, arrayAddress(a, offset),
             this, a, offset,
             (arr, off, v)
-            -> v.stOp(arr, off,
+            -> v.stOp(arr, (int) off,
                       (arr_, off_, i, e) -> arr_[off_+i] = e));
     }
 
@@ -3703,7 +3703,7 @@ public abstract class LongVector extends AbstractVector<Long> {
             a, arrayAddress(a, offset),
             this, m, a, offset,
             (arr, off, v, vm)
-            -> v.stOp(arr, off, vm,
+            -> v.stOp(arr, (int) off, vm,
                       (arr_, off_, i, e) -> arr_[off_ + i] = e));
     }
 
@@ -3773,7 +3773,7 @@ public abstract class LongVector extends AbstractVector<Long> {
             this, a, offset,
             (arr, off, v) -> {
                 ByteBuffer wb = wrapper(arr, NATIVE_ENDIAN);
-                v.stOp(wb, off,
+                v.stOp(wb, (int) off,
                         (tb_, o, i, e) -> tb_.putLong(o + i * 8, e));
             });
     }
@@ -3792,7 +3792,7 @@ public abstract class LongVector extends AbstractVector<Long> {
             this, m, a, offset,
             (arr, off, v, vm) -> {
                 ByteBuffer wb = wrapper(arr, NATIVE_ENDIAN);
-                v.stOp(wb, off, vm,
+                v.stOp(wb, (int) off, vm,
                         (tb_, o, i, e) -> tb_.putLong(o + i * 8, e));
             });
     }
@@ -3806,7 +3806,7 @@ public abstract class LongVector extends AbstractVector<Long> {
                 this, bb, offset,
                 (buf, off, v) -> {
                     ByteBuffer wb = wrapper(buf, NATIVE_ENDIAN);
-                    v.stOp(wb, off,
+                    v.stOp(wb, (int) off,
                             (wb_, o, i, e) -> wb_.putLong(o + i * 8, e));
                 });
     }
@@ -3824,7 +3824,7 @@ public abstract class LongVector extends AbstractVector<Long> {
                 this, m, bb, offset,
                 (buf, off, v, vm) -> {
                     ByteBuffer wb = wrapper(buf, NATIVE_ENDIAN);
-                    v.stOp(wb, off, vm,
+                    v.stOp(wb, (int) off, vm,
                             (wb_, o, i, e) -> wb_.putLong(o + i * 8, e));
                 });
     }
@@ -3836,10 +3836,10 @@ public abstract class LongVector extends AbstractVector<Long> {
         ScopedMemoryAccess.storeIntoMemorySegment(
                 vsp.vectorType(), vsp.elementType(), vsp.laneCount(),
                 this,
-                (MemorySegmentProxy) ms, (int) offset, // @@@ downcast from long to int
+                (MemorySegmentProxy) ms, offset,
                 (msp, off, v) -> {
                     var layout = ValueLayout.JAVA_LONG.withBitAlignment(8);
-                    v.stOp((MemorySegment) msp, off,
+                    v.stOp((MemorySegment) msp, (int) off, // @@@ downcast from long to int
                             (ms_, o, i, e) -> ms_.set(layout, o + i * 8L, e));
                 });
     }
@@ -3855,10 +3855,10 @@ public abstract class LongVector extends AbstractVector<Long> {
         ScopedMemoryAccess.storeIntoMemorySegmentMasked(
                 vsp.vectorType(), maskClass, vsp.elementType(), vsp.laneCount(),
                 this, m,
-                (MemorySegmentProxy) ms, (int) offset, // @@@ downcast from long to int
+                (MemorySegmentProxy) ms, offset,
                 (msp, off, v, vm) -> {
                     var layout = ValueLayout.JAVA_LONG.withBitAlignment(8);
-                    v.stOp((MemorySegment) msp, off, vm,
+                    v.stOp((MemorySegment) msp, (int) off, vm,  // @@@ downcast from long to int
                             (ms_, o, i, e) -> ms_.set(layout, o + i * 8L, e));
                 });
     }
