+++ src/jdk.incubator.vector/share/classes/jdk/incubator/vector/VectorMask.java
@@ -210,7 +210,7 @@ public abstract class VectorMask<E> extends jdk.internal.vm.vector.VectorSupport
                 bits, (long) offset + Unsafe.ARRAY_BOOLEAN_BASE_OFFSET,
                 bits, offset, vsp,
                 (c, idx, s)
-                  -> s.opm(n -> c[idx + n]));
+                  -> s.opm(n -> c[((int )idx) + n]));
     }
 
     /**
